// src/modules/Access/enums.ts
export enum RoleStatus {
  ENABLED = 1,
  DISABLED = 0,
}

/**
 * 权限类型
 * @enum FOLDER: 文件夹
 * @enum MENU: 目录
 * @enum BUTTON: 按钮
 * */
export enum AccessType {
  FOLDER = 0,
  MENU = 1,
  BUTTON = 2,
}

/**
 * 权限状态
 * @enum ENABLED: 启用
 * @enum DISABLED: 禁用
 * */
export enum AccessStatus {
  ENABLED = 1,
  DISABLED = 0,
}